Rename: Show.jsx를 components 파일로 이동
