+++ src/pages/Mypage.jsx
@@ -1,0 +1,7 @@
+import React from 'react';
+
+function Mypage() {
+  return <div>Mypage</div>;
+}
+
+export default Mypage;
